pyproject.toml
[tool.poetry]
name = "nest"
version = "0.1.0"
description = "Nest is a project for managing my home (nest)"
authors = ["Daniel Kjellid <daniel.kjellid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10.5"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.46"
alembic = "^1.9.2"
fastapi-utils = "^0.2.1"
asyncpg = "^0.27.0"
psycopg2 = "^2.9.5"
greenlet = "^2.0.1"
ipython = "^8.8.0"
structlog = "^22.3.0"
asgi-correlation-id = "^3.2.1"
pyjwt = "^2.6.0"
passlib = "^1.7.4"
httpx = "^0.23.3"
databases = "^0.7.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
ruff = "^0.0.236"
mypy = "^0.991"
pytest = "^7.2.1"
pytest-pretty = "^1.0.1"
pytest-env = "^0.8.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pytest-lazy-fixture = "^0.6.3"
pytest-xdist = "^3.1.0"
pytest-asyncio = "^0.20.3"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#########
# Black #
#########

[tool.black]
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = 'revisions'
safe = true

########
# Mypy #
########

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
follow_imports = "normal"
ignore_missing_imports = true
strict = true

disallow_subclassing_any = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

########
# Ruff #
########

[tool.ruff]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "revisions",
]
select = [
    # Regular flake8 rules
    "C", "E", "F", "W",
    # flake8-bugbear rules
    "B",
    "I",
    # Pylint rules
    "PLC", "PLE", "PLR", "PLW",
     "RUF",
]
ignore = [
    # C408: Unnecessary <dict/list/tuple> call - rewrite as a literal
    "C408",
    # B905: zip() without explicit strict,
    "B905",
    # E501: Leave line-length stuff for black to deal with
    "E501",
    # Disable magic value comparison. They're perfectly valid in tests and quite a few
    "PLR2004",
]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"

##########
# Pytest #
##########

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["nest"]
python_files = "test_*.py"
# Include captured log messages in system-out in CI report file
junit_logging = "system-out"
# --- pytest-env settings
env = ["ENV_PATH=.env.test"]
